sdu-hole/requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.6
sqlalchemy[asyncio]==2.0.35
asyncpg==0.30.0
aiosqlite==0.20.0
pydantic-settings==2.5.2
python-jose[cryptography]==3.3.0
aiosmtplib==3.0.2
python-multipart==0.0.12
passlib[bcrypt]==1.7.4
